# Conflicts:
#	README.md
#	extensions/chromium/.eslintrc
#	extensions/chromium/background.js
#	extensions/chromium/options/options.html
#	extensions/chromium/pdfHandler.js
#	extensions/chromium/restoretab.html
#	external/jbig2/LICENSE_PDFJS_JBIG2
#	l10n/be/viewer.ftl
#	l10n/bg/viewer.ftl
#	l10n/cs/viewer.ftl
#	l10n/da/viewer.ftl
#	l10n/dsb/viewer.ftl
#	l10n/el/viewer.ftl
#	l10n/en-GB/viewer.ftl
#	l10n/eo/viewer.ftl
#	l10n/es-AR/viewer.ftl
#	l10n/es-CL/viewer.ftl
#	l10n/fr/viewer.ftl
#	l10n/fy-NL/viewer.ftl
#	l10n/he/viewer.ftl
#	l10n/hsb/viewer.ftl
#	l10n/hu/viewer.ftl
#	l10n/ia/viewer.ftl
#	l10n/is/viewer.ftl
#	l10n/it/viewer.ftl
#	l10n/kk/viewer.ftl
#	l10n/ko/viewer.ftl
#	l10n/nl/viewer.ftl
#	l10n/nn-NO/viewer.ftl
#	l10n/pl/viewer.ftl
#	l10n/pt-BR/viewer.ftl
#	l10n/rm/viewer.ftl
#	l10n/ru/viewer.ftl
#	l10n/sc/viewer.ftl
#	l10n/sk/viewer.ftl
#	l10n/skr/viewer.ftl
#	l10n/sl/viewer.ftl
#	l10n/sq/viewer.ftl
#	l10n/sv-SE/viewer.ftl
#	l10n/tg/viewer.ftl
#	l10n/th/viewer.ftl
#	l10n/tr/viewer.ftl
#	l10n/uk/viewer.ftl
#	l10n/vi/viewer.ftl
#	l10n/zh-CN/viewer.ftl
#	l10n/zh-TW/viewer.ftl
#	package-lock.json
#	package.json
#	pdfjs.config
#	src/core/annotation.js
#	src/core/struct_tree.js
#	src/display/annotation_layer.js
#	src/display/display_utils.js
#	src/display/draw_layer.js
#	src/display/editor/alt_text.js
#	src/display/editor/editor.js
#	src/display/editor/highlight.js
#	src/display/editor/stamp.js
#	src/display/editor/toolbar.js
#	src/display/fetch_stream.js
#	src/display/network.js
#	src/display/network_utils.js
#	src/display/node_stream.js
#	test/integration/accessibility_spec.mjs
#	test/integration/highlight_editor_spec.mjs
#	test/integration/stamp_editor_spec.mjs
#	test/integration/test_utils.mjs
#	test/integration/viewer_spec.mjs
#	test/pdfs/.gitignore
#	test/test_manifest.json
#	test/unit/network_utils_spec.js
#	test/unit/struct_tree_spec.js
#	web/annotation_editor_layer_builder.css
#	web/annotation_layer_builder.js
#	web/app.js
#	web/new_alt_text_manager.js
#	web/pdf_page_view.js
#	web/pdf_viewer.css
#	web/struct_tree_layer_builder.js
#	web/viewer.css

+++ src/display/annotation_storage.js
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-import { objectFromMap, shadow, unreachable } from "../shared/util.js";
+import { shadow, unreachable } from "../shared/util.js";
 import { AnnotationEditor } from "./editor/editor.js";
 import { MurmurHash3_64 } from "../shared/murmurhash3.js";
 
@@ -31,6 +31,8 @@ class AnnotationStorage {
 
   #modifiedIds = null;
 
+  #editorsMap = null;
+
   #storage = new Map();
 
   constructor() {
@@ -41,6 +43,17 @@ class AnnotationStorage {
     this.onSetModified = null;
     this.onResetModified = null;
     this.onAnnotationEditor = null;
+
+    if (typeof PDFJSDev === "undefined" || PDFJSDev.test("TESTING")) {
+      // For testing purposes.
+      Object.defineProperty(this, "_setValues", {
+        value: obj => {
+          for (const [key, val] of Object.entries(obj)) {
+            this.setValue(key, val);
+          }
+        },
+      });
+    }
   }
 
   /**
@@ -72,6 +85,13 @@ class AnnotationStorage {
    * @param {string} key
    */
   remove(key) {
+    const storedValue = this.#storage.get(key);
+    if (storedValue === undefined) {
+      return;
+    }
+    if (storedValue instanceof AnnotationEditor) {
+      this.#editorsMap.delete(storedValue.annotationElementId);
+    }
     this.#storage.delete(key);
 
     if (this.#storage.size === 0) {
@@ -111,11 +131,11 @@ class AnnotationStorage {
       this.#setModified();
     }
 
-    if (
-      value instanceof AnnotationEditor &&
-      typeof this.onAnnotationEditor === "function"
-    ) {
-      this.onAnnotationEditor(value.constructor._type);
+    if (value instanceof AnnotationEditor) {
+      (this.#editorsMap ||= new Map()).set(value.annotationElementId, value);
+      if (typeof this.onAnnotationEditor === "function") {
+        this.onAnnotationEditor(value.constructor._type);
+      }
     }
   }
 
@@ -126,22 +146,6 @@ class AnnotationStorage {
    */
   has(key) {
     return this.#storage.has(key);
-  }
-
-  /**
-   * @returns {Object | null}
-   */
-  getAll() {
-    return this.#storage.size > 0 ? objectFromMap(this.#storage) : null;
-  }
-
-  /**
-   * @param {Object} obj
-   */
-  setAll(obj) {
-    for (const [key, val] of Object.entries(obj)) {
-      this.setValue(key, val);
-    }
   }
 
   get size() {
@@ -218,9 +222,23 @@ class AnnotationStorage {
   get editorStats() {
     let stats = null;
     const typeToEditor = new Map();
+    let numberOfEditedComments = 0;
+    let numberOfDeletedComments = 0;
     for (const value of this.#storage.values()) {
       if (!(value instanceof AnnotationEditor)) {
+        if (value.popup) {
+          if (value.popup.deleted) {
+            numberOfDeletedComments += 1;
+          } else {
+            numberOfEditedComments += 1;
+          }
+        }
         continue;
+      }
+      if (value.isCommentDeleted) {
+        numberOfDeletedComments += 1;
+      } else if (value.hasEditedComment) {
+        numberOfEditedComments += 1;
       }
       const editorStats = value.telemetryFinalData;
       if (!editorStats) {
@@ -245,6 +263,16 @@ class AnnotationStorage {
         counters.set(val, count + 1);
       }
     }
+    if (numberOfDeletedComments > 0 || numberOfEditedComments > 0) {
+      stats ||= Object.create(null);
+      stats.comments = {
+        deleted: numberOfDeletedComments,
+        edited: numberOfEditedComments,
+      };
+    }
+    if (!stats) {
+      return null;
+    }
     for (const [type, editor] of typeToEditor) {
       stats[type] = editor.computeTelemetryFinalData(stats[type]);
     }
@@ -255,6 +283,19 @@ class AnnotationStorage {
     this.#modifiedIds = null;
   }
 
+  updateEditor(annotationId, data) {
+    const value = this.#editorsMap?.get(annotationId);
+    if (value) {
+      value.updateFromAnnotationLayer(data);
+      return true;
+    }
+    return false;
+  }
+
+  getEditor(annotationId) {
+    return this.#editorsMap?.get(annotationId) || null;
+  }
+
   /**
    * @returns {{ids: Set<string>, hash: string}}
    */
@@ -263,20 +304,22 @@ class AnnotationStorage {
       return this.#modifiedIds;
     }
     const ids = [];
-    for (const value of this.#storage.values()) {
-      if (
-        !(value instanceof AnnotationEditor) ||
-        !value.annotationElementId ||
-        !value.serialize()
-      ) {
-        continue;
+    if (this.#editorsMap) {
+      for (const value of this.#editorsMap.values()) {
+        if (!value.serialize()) {
+          continue;
+        }
+        ids.push(value.annotationElementId);
       }
-      ids.push(value.annotationElementId);
     }
     return (this.#modifiedIds = {
       ids: new Set(ids),
       hash: ids.join(","),
     });
+  }
+
+  [Symbol.iterator]() {
+    return this.#storage.entries();
   }
 }
